refactor: sidebar navigation
* split components to different files and created a folder component for them.
* renaming Navigation component to SidebarNavigation
